+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Academies/PupilNumbers.cshtml.cs
@@ -1,8 +1,8 @@
 using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Data.Enums;
-using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
+using DfE.FindInformationAcademiesTrusts.Services.Export;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using Microsoft.AspNetCore.Mvc;
 
@@ -13,9 +13,9 @@ public class PupilNumbersModel : AcademiesPageModel
     public IAcademyService AcademyService { get; }
     public AcademyPupilNumbersServiceModel[] Academies { get; set; } = default!;
 
-    public PupilNumbersModel(ITrustProvider trustProvider, IDataSourceService dataSourceService,
+    public PupilNumbersModel(IDataSourceService dataSourceService,
         ILogger<PupilNumbersModel> logger, ITrustService trustService, IAcademyService academyService, IExportService exportService, IDateTimeProvider dateTimeProvider)
-        : base(trustProvider, dataSourceService, trustService, academyService, exportService, logger, dateTimeProvider)
+        : base(dataSourceService, trustService, exportService, logger, dateTimeProvider)
     {
         AcademyService = academyService;
         PageTitle = "Academies pupil numbers";