+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Academies/AcademiesPageModelTests.cs
@@ -1,9 +1,9 @@
 ﻿using DfE.FindInformationAcademiesTrusts.Data;
 using DfE.FindInformationAcademiesTrusts.Pages;
 using DfE.FindInformationAcademiesTrusts.Pages.Trusts.Academies;
-using DfE.FindInformationAcademiesTrusts.Services;
 using DfE.FindInformationAcademiesTrusts.Services.Academy;
 using DfE.FindInformationAcademiesTrusts.Services.DataSource;
+using DfE.FindInformationAcademiesTrusts.Services.Export;
 using DfE.FindInformationAcademiesTrusts.Services.Trust;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -12,7 +12,6 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Academies
 {
     public class AcademiesPageModelTests
     {
-        private readonly Mock<ITrustProvider> _mockTrustProvider = new();
         private readonly Mock<ITrustService> _mockTrustService = new();
         private readonly Mock<IAcademyService> _mockAcademyService = new();
         private readonly Mock<IExportService> _mockExportService = new();
@@ -24,7 +23,7 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Academies
 
         public AcademiesPageModelTests()
         {
-            _sut = new AcademiesDetailsModel(_mockTrustProvider.Object, _mockDataSourceService.Object, _mockOtherServicesLinkBuilder.Object, _mockLogger.Object, _mockTrustService.Object, _mockAcademyService.Object, _mockExportService.Object, _mockDateTimeProvider.Object);
+            _sut = new AcademiesDetailsModel(_mockDataSourceService.Object, _mockOtherServicesLinkBuilder.Object, _mockLogger.Object, _mockTrustService.Object, _mockAcademyService.Object, _mockExportService.Object, _mockDateTimeProvider.Object);
         }
 
         [Fact]
@@ -32,27 +31,13 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Academies
         {
             // Arrange
             string uid = "1234";
-            var trust = new Trust("1234",
-                                  "Sample Trust",
-                                  "1001",
-                                  "12345",
-                                  "Multi-academy trust",
-                                  "Address",
-                                  DateTime.Now,
-                                  "123456",
-                                  "Region",
-                                  [],
-                                  [],
-                                  null,
-                                  null,
-                                  "Open");
-            var trustSummary = new TrustSummaryServiceModel(uid, "Sample Trust", "Multi-academy trust", 0);
-            var expectedBytes = new byte[] { 1, 2, 3 };
-            var ofstedRatings = Array.Empty<AcademyOfstedServiceModel>();
 
-            _mockTrustProvider.Setup(x => x.GetTrustByUidAsync(uid)).ReturnsAsync(trust);
+            var trustSummary = new TrustSummaryServiceModel(uid, "Sample Trust", "Multi-academy trust", 0);
+            byte[] expectedBytes = [1, 2, 3];
+
             _mockTrustService.Setup(x => x.GetTrustSummaryAsync(uid)).ReturnsAsync(trustSummary);
-            _mockExportService.Setup(x => x.ExportAcademiesToSpreadsheetUsingProvider(trust, trustSummary, ofstedRatings)).Returns(expectedBytes);
+            _mockExportService.Setup(x => x.ExportAcademiesToSpreadsheetAsync(uid)).ReturnsAsync(expectedBytes);
+
 
             // Act
             var result = await _sut.OnGetExportAsync(uid);
@@ -69,7 +54,6 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Academies
             // Arrange
             string uid = "invalid-uid";
 
-            _mockTrustProvider.Setup(x => x.GetTrustByUidAsync(uid)).ReturnsAsync((Trust?)null);
             _mockTrustService.Setup(x => x.GetTrustSummaryAsync(uid)).ReturnsAsync((TrustSummaryServiceModel?)null);
 
             // Act
@@ -84,27 +68,11 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Academies
         {
             // Arrange
             string uid = "1234";
-            var trustWithIllegalChars = new Trust("1234",
-                                                  "Sample/Trust:Name?", // Trust name with illegal characters
-                                                  "1001",
-                                                  "12345",
-                                                  "Multi-academy trust",
-                                                  "Address",
-                                                  DateTime.Now,
-                                                  "123456",
-                                                  "Region",
-                                                  [],
-                                                  [],
-                                                  null,
-                                                  null,
-                                                  "Open");
             var trustSummary = new TrustSummaryServiceModel(uid, "Sample/Trust:Name?", "Multi-academy trust", 0);
             var expectedBytes = new byte[] { 1, 2, 3 };
-            var ofstedRatings = Array.Empty<AcademyOfstedServiceModel>();
 
-            _mockTrustProvider.Setup(x => x.GetTrustByUidAsync(uid)).ReturnsAsync(trustWithIllegalChars);
             _mockTrustService.Setup(x => x.GetTrustSummaryAsync(uid)).ReturnsAsync(trustSummary);
-            _mockExportService.Setup(x => x.ExportAcademiesToSpreadsheetUsingProvider(trustWithIllegalChars, trustSummary, ofstedRatings)).Returns(expectedBytes);
+            _mockExportService.Setup(x => x.ExportAcademiesToSpreadsheetAsync(uid)).ReturnsAsync(expectedBytes);
 
             // Act
             var result = await _sut.OnGetExportAsync(uid);