+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Contacts/EditSfsoLeadModelTests.cs
@@ -11,7 +11,6 @@ namespace DfE.FindInformationAcademiesTrusts.UnitTests.Pages.Trusts.Contacts;
 public class EditSfsoLeadModelTests
 {
     private readonly EditSfsoLeadModel _sut;
-    private readonly Mock<ITrustProvider> _mockTrustProvider = new();
 
     private readonly MockDataSourceService _mockDataSourceService = new();
     private readonly Mock<ITrustService> _mockTrustService = new();
@@ -28,9 +27,9 @@ public class EditSfsoLeadModelTests
         _mockTrustService.Setup(t => t.GetTrustSummaryAsync(_fakeTrust.Uid))
             .ReturnsAsync(_fakeTrust);
 
-        _sut = new EditSfsoLeadModel(_mockTrustProvider.Object, _mockDataSourceService.Object,
+        _sut = new EditSfsoLeadModel(_mockDataSourceService.Object,
                 new MockLogger<EditSfsoLeadModel>().Object, _mockTrustService.Object)
-            { Uid = "1234" };
+        { Uid = "1234" };
     }
 
     [Fact]